pyproject.toml
[tool.black]
line-length = 119
target-version = ['py310']

[tool.poetry]
name = "wraeclast-whisperer"
version = "0.0.1"
description = "An Agentic Path of Exile Assistant"
authors = ["darecstowell"]
readme = "README.md"

packages = [
    { include = "wraeclast_whisperer", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.12"
# jinja2 = "*"
openai = "^1.57.4"
panel = "^1.5.4"
pydantic = "^2.10.3"
# pyyaml = "*"
requests = "*"
pymediawiki = "^0.7.4"
python-dotenv = "^1.0.1"
